Moving readmodel adapters to shared models so that they can be used in tests
