+++ packages/catalog-readmodel-writer/src/consumerServiceV2.ts
@@ -1,7 +1,10 @@
 import { EServiceCollection, logger } from "pagopa-interop-commons";
-import { EServiceEventEnvelopeV2, fromEServiceV2 } from "pagopa-interop-models";
+import {
+  EServiceEventEnvelopeV2,
+  fromEServiceV2,
+  toReadModelEService,
+} from "pagopa-interop-models";
 import { match } from "ts-pattern";
-import { toReadModelEService } from "./model/legacy/eserviceAdapter.js";
 
 export async function handleMessageV2(
   message: EServiceEventEnvelopeV2,
