+++ packages/models/src/index.ts
@@ -7,15 +7,16 @@ export * from "./agreement/agreementEvents.js";
 export * from "./agreement/protobufConverter.js";
 
 export * from "./attribute/attribute.js";
-export * from "./attribute/protobufConverter.js";
 export * from "./attribute/attributeEvents.js";
+export * from "./attribute/protobufConverter.js";
 
 export * from "./email/email.js";
 
 export * from "./eservice/eservice.js";
-export * from "./eservice/eserviceEvents.js";
+export * from "./eservice/eserviceAdapter.js";
 export * from "./eservice/protobufConverterV1.js";
 export * from "./eservice/protobufConverterV2.js";
+export * from "./eservice/eserviceEvents.js";
 
 export * from "./institution/institution.js";
 
@@ -32,8 +33,8 @@ export * from "./protobuf/protobuf.js";
 export * from "./protobuf/utils.js";
 
 // Read models
-export * from "./readModels/eserviceReadModel.js";
-export * from "./readModels/readModels.js";
+export * from "./read-models/eserviceReadModel.js";
+export * from "./read-models/readModels.js";
 
 // Utilities
 export * from "./brandedIds.js";
